Explicitly show image date and be able to override it
I noticed that when Dan sent me an image, for some reason the date
wasn't in the metadata and my app was saving a file called
null/null/null-null-null...

This will now make it obvious what date will be used, and allow me to
change the date if I want to. It also prevents me from saving if there
is no date.